qei driver: fix potential overflow in qeiUpdateI
when quiUpdate isn't called for while an cnt(uint16_t) is above 32767
delta(int16_t) can overflow.

+++ os/hal/ports/STM32/LLD/TIMv1/hal_qei_lld.h
@@ -238,7 +238,7 @@ typedef uint16_t qeicnt_t;
 /**
  * @brief   QEI delta type.
  */
-typedef int16_t qeidelta_t;
+typedef int32_t qeidelta_t;
 
 /**
  * @brief   Driver configuration structure.
